refactor: extract KeyMetricsQuery from KeyCountStore (Close #225)
